langgraph-agents/requirements.txt
# LangGraph Multi-Agent System Dependencies

# Core LangGraph framework
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0

# State management and persistence
redis>=4.5.0
pydantic>=2.0.0

# AI/ML dependencies
openai>=1.0.0
anthropic>=0.8.0
tiktoken>=0.5.0

# Financial API integrations
alpha-vantage>=2.3.1
yfinance>=0.2.18
requests>=2.31.0

# LibreOffice integration
uno  # Will be available in LibreOffice Python environment
pyuno  # LibreOffice Python UNO bridge

# Performance and monitoring
psutil>=5.9.0
prometheus-client>=0.16.0

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Development utilities
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.3.0

# Security and credentials
cryptography>=41.0.0
keyring>=24.0.0
python-dotenv>=1.0.0

# Async processing
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0